New free user

+++ app/Http/Controllers/Auth/RegisterController.php
@@ -3,13 +3,15 @@
 namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
+use App\Models\UserProfile;
 use App\User;
 use Carbon\Carbon;
+use Cart;
+use Illuminate\Auth\Events\Registered;
 use Illuminate\Foundation\Auth\RegistersUsers;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
-use Illuminate\Http\Request;
-use Cart;
 
 class RegisterController extends Controller
 {
@@ -55,7 +57,7 @@ class RegisterController extends Controller
             'name' => 'required|max:255',
             'email' => 'required|email|max:255|unique:users',
             'password' => 'required|min:6|confirmed',
-            'plan' => 'required|in:1,2,3',
+            'plan' => 'required|in:0,1,2,3',
         ]);
     }
 
@@ -63,37 +65,60 @@ class RegisterController extends Controller
     {
         $this->validator($request->all())->validate();
 
-        $request->session()->put('email_register', $request->input('name'));
-        $request->session()->put('name_register', $request->input('email'));
-        $request->session()->put('pass_register', $request->input('password'));
-        $request->session()->put('plan_register', $request->input('plan'));
-        
-        Cart::destroy();
-        
-        switch ($request->input('plan')) {
-            case '1':
-                Cart::add('plan 1', 'plan 1 for 1 years', 1, 25);
-                $years = 1;
-                break;
-            case '2':
-                Cart::add('plan 2', 'plan 2 for 2 years', 1, 45);
-                $years = 2;
-                break;
-            case '3':
-                Cart::add('plan 3', 'plan 3 for 5 years', 1, 95);
-                $years = 5;
-                break;
-            default:
-                $years = 0;
-                break;
-        }
+        if ($request->input('plan') != 0) {
+            $request->session()->put('email_register', $request->input('name'));
+            $request->session()->put('name_register', $request->input('email'));
+            $request->session()->put('pass_register', $request->input('password'));
+            $request->session()->put('plan_register', $request->input('plan'));
+            
+            Cart::destroy();
+            
+            switch ($request->input('plan')) {
+                case '1':
+                    Cart::add('plan 1', 'plan 1 for 1 years', 1, 25);
+                    $years = 1;
+                    break;
+                case '2':
+                    Cart::add('plan 2', 'plan 2 for 2 years', 1, 45);
+                    $years = 2;
+                    break;
+                case '3':
+                    Cart::add('plan 3', 'plan 3 for 5 years', 1, 95);
+                    $years = 5;
+                    break;
+                default:
+                    $years = 0;
+                    break;
+            }
 
-        return redirect()->route('payment.register');
-        
+            return redirect()->route('payment.register');
+        } else {
+                $this->validator($request->all())->validate();
+
+                event(new Registered($user = $this->create($request->all())));
+
+                $this->guard()->login($user);
+
+                return $this->registered($request, $user)
+                            ?: redirect($this->redirectPath());
+        }        
     }
 
     protected function create(array $data)
     {
+        $user = User::create([
+            'name' => $data['name'],
+            'email' => $data['email'],
+            'password' => bcrypt($data['password']),
+        ]);
+
+        $profile = new UserProfile;
+        $profile->plan = 0;
+        $profile->user_id = $user->id;
+        $profile->save();
+
+        return $user;
+
         // DB::beginTransaction();
         // try {
         //     $user = User::create([
